Rename transformer folder
